Use joinpsbts in name_ant_workflow.py.
This simplifies the name_ant_workflow.py test / example code (for how to
do atomic name trades using PSBTs).  In particular, we do not yet add the
name output when funding the transaction with the wallet, to avoid a bug
with the total balances.  Also, we use joinpsbts to construct the combined
transaction, rather than doing it manually.