Strip name prefixes in descriptor wallet.
Make sure that the descriptor wallet properly works with name
transactions, which mainly requires us to strip the name prefix before
we attempt to look up the matching key.

Also updates the regtest driver to run some of the main name tests
(which involve various wallet operations) with a descriptor wallet
in addition to a legacy one.