Avoid legacy SPK for deterministic salts.
When computing the deterministic salts for name registrations, use
generic methods to extract the SigningProvider instead of explicitly
relying on GetLegacyScriptPubKeyMan (which doesn't work for descriptor
wallets).

For this to work, we need to expand the ScriptPubKeyMan interface with
a new method to extract the SigningProvider with private keys, which
is not directly possible with the upstream interface.