Move IsBurn to script/standard.
This moves the IsBurn method, which checks if a script is a CHI burn
according to how moves are supposed to do burns, to script/standard
(from out of being a private method in zmqgames).

In the future, we want to expose this data not just through ZMQ but
also through raw-transaction decoding.