Switch default value encoding to UTF-8.
Values encoded as UTF-8 JSON are valid according to the consensus
rules, so it makes sense to use UTF-8 as the default encoding for
values in name_show and related methods.

With this, names and values will be default-encoded in UTF-8, which
means that all data that is valid according to consensus rules can
actually be represented correctly in those RPC results.