Update name_ant_workflow.py.
Upstream https://github.com/bitcoin/bitcoin/pull/17211 solved a
long-standing issue for doing atomic trades with names (namely that
the buyer's wallet was not able to directly fund a transaction with the
non-wallet name input in it).

Based on this, we can now streamline name_ant_workflow.py to make use
of this new feature.  This allows us to properly use fundrawtransaction
to fund the final transaction including already the name input and output,
and thus with correct fee estimation done.  (Previously we just funded a
dummy transaction with a guessed higher fee, and then added in the name
input and output manually, hoping the fee would still be fine.)