Show times for fee estimate in blocks only.
In the fee estimation UI, only show number of blocks and not estimated
times.  The times need the block spacing, which in Xaya depends on the
current block height; and this leads to a problem with initialisation
order.

This needs the client model set, but it would need it already when
constructing a WalletView, while the client model is only set
on the WalletView after construction.