Script verify flag for long name salts.
This adds a new verify flag SCRIPT_VERIFY_NAMES_LONG_SALT, which (when
turned on) requires salts in name_firstupdate operations to be exactly
20 bytes long.

Old versions of Namecoin were using short and thus insecure salts, while
the default behaviour in Namecoin Core has been usage of full-length
salts for a long time now.  Let's make sure all future users are not
using insecure salts anymore.

For now, this new flag is turned on in the policy but not turned on
for new blocks (i.e. this is not yet a fork).