Some fixes for signet.
This includes some adjustments to the signet and miner:  Specify testnet
base58 prefixes for signet as well on Xaya, set the powdata.xayahash
argument in the miner (so we can neoscrypt hash from Python) and
actually set the powdata nBits.  Also update "xaya-util grind" to
work with pure block headers and SHA-256d explicitly.

It is not yet fully functional, though, as the difficulty updates
and estimation is not working properly for the test yet.