No coin selection if provided input suffices.
In Namecoin, the wallet may sometimes get a pre-defined input for
constructing transactions (the name input).  If its value is large
enough to actually support the entire transaction (because it was more
than the minimum 0.01 NMC), we can entirely avoid coin selection.

This fixes a potential issue introduced with recent upstream
coin-selection changes, if the set of selected coins is empty.