Update anti-DoS headers sync for auxpow.
Update the new "anti DoS headers sync" from upstream's
https://github.com/bitcoin/bitcoin/pull/25717 for auxpow.  The main change
is that we need to apply the "max size" logic for header messages
also in this mode, and that we need to make sure we get full block
headers (including auxpow field) during redownload phase so we can
save them properly to permanent storage.

The latter also breaks Bitcoin's memory analysis in the auxpow case,
but for now we are fine with that and just go with the parameters from
upstream and see how it works.