Remove xaya-hash binary.
The xaya-hash binary was a hack, so we could use the C++ neoscrypt
hashing code for Python regtests.  Since quite some time, we now have
a working Python package for neoscrypt, so it makes more sense to use that
in regtests instead.  The xaya-hash binary can be removed completely,
which also eases maintenance.