Use plain block hash for rngseed.
This schedules a "fork" (although it only affects nodes running GSPs and
not the main network at all), which changes the rngseed to be based on
the block hash instead of PoW.

While the PoW is desirable because it is harder to manipulate by miners,
it is not part of blockchain consensus, and thus it is not guaranteed
that random numbers are the same for everyone in the network.

This will take effect on October 15th, 2023.