Use separate field for name op in CRecipient.
Upstream changed CRecipient to hold a CTxDestination for the
"recipient address" instead of a raw script, see
https://github.com/bitcoin/bitcoin/pull/28246.

While this continues to "just work" for name operations since it
still supports arbitrary scripts, this is of course not fully in
line with the upstream refactor.

In this commit, we change CRecipient to hold a name operation, if any,
in a separate field that just contains the name script.  The recipient
address is kept in the CTxDestination.

With this, name operations will more transparently blend in with the
upstream code, and will more directly benefit from future things like
silent payments.