Fix getheaders for blocks not on disk yet.
Unlike upstream Bitcoin, it is impossible in Namecoin and Xaya to
convert a CBlockIndex directly to a block header; for this, we need
to retrieve the block from disk.

With this, it may happen that a node tries (and fails to) load blocks
from disk it does not have yet while serving getheaders, e.g. when still
syncing with assumeutxo.

In this commit, we make it properly catch this situation, and respond
by ignoring the request.